requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
GDAL==3.6.3
psycopg2-binary==2.9.5
pytz==2023.2
sqlparse==0.4.3
